analysis/requirements.txt
numpy
pandas
google-cloud-storage
researchpy
statsmodels
